requirements.txt
click==8.1.7
Flask==2.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pymongo==3.12.0
python-dotenv==1.0.1
Werkzeug==2.0.3
Flask-Login==0.6.2
bcrypt==4.0.1
